Fix external program directory name issue
When running an external program, $d can be used to pass the
directory name of the file being edited. This includes a trailing
slash (backslash on Windows). If the name contains spaces, then
it is surrounded by double quotes. However, a trailing backslash
followed by a double quote escapes the quote. This meant a
program such as ebookmaker which requires an output-dir
argument could not be run from within Guiguts on a file with
spaces in its pathname.
This edit provides an alternative $D which is identical to $d
but with the trailing slash removed.