Fix Windows problem with spaces in dir name
When a process is spawned using "start", e.g. View in
Browser from External Operations, or View Project
Comments, it was failing on Window if the path had
spaces in it.
This is due to the start command in Windows
interpreting the first argument as the title of the
window if it has quotes round it. Solved by adding
a dummy title with spaces to force it to be quoted.