Fix wrong projectID loading a new file
If a currently loaded file has a project ID, then a
new file is loaded that has no .bin file, the project
ID from the previous file did not get cleared. This
meant that links to project forum, etc., were
wrong.

Project ID is now cleared when file is loaded.