Fix single-clicking in error windows
In the bookloupe, jeebies and "others" error
windows, single-clicking on an error was
supposed to take you to the error in the main
window. It actually took you to the previously
selected error.

This was due to the event being triggered on
the mouse button being pressed, rather than
released, so the list of errors had not actually
selected the one clicked on yet.

This should also fix #123, that clicking on errors
after running pphtml sometimes caused the bell
(system beep). When the error does not have a
line number, GG searches for the text at the start
of the line (between plus and colon symbol). The
error can be clicked multiple times to search
through the occurrences, and then the bell is
sounded when there are no more search
matches, just as with a normal search. Because
of the "previous error" bug, if the user removed
an unused definition, e.g. "h6", then clicked on
their next error (unused h5), it would first do a
search for h6 (causing a bell) before the user clicked again on the h5 error to jump to that.