Stop poetry in footnotes being enclosed in paragraph markup
The footnote code made its own attempt to add `<p>` and `</p>` markup, but didn't
have the logic which is in the main HTML conversion code.

The footnote code now doesn't attempt to start/end paragraphs - it leaves it to the
main code, which has been slightly tweaked. It previously avoided adding `<p>`
immediately before the start of a div, heading or paragraph, but didn't avoid adding
it before the end of these. Also its check for avoiding an hr or br element was too
restrictive.

Fixes #101