Make jeebies treat file without he/be as odd but not an error
If a file had "he" but not "be" or vice versa, jeebies reported it as odd, but didn't
return an error status. If a file had neither "he" nor "be", it reported it as odd,
and also exited with an error status. This caused GG to say it had been unable to
run jeebies.

Jeebies now exits with status 0 in this case, and the user see the "odd" file message
and that there are no queries.

Only the Windows exe has been rebuilt.

Fixes #147