Make Variation Selectors Unicode block available
This makes the Variation Selectors block available in the Unicode menu and in the
selection menu in the Unicode block dialog. It also makes the Variation Selectors
available for searching in the Unicode Character Search dialog.

Note that some of these are typically not visible in a font, but the apparently empty
box can still be selected to insert them into the text.

Number of available blocks is too large for a 2-column menu to fit on many screens
so now forced to a 3-column menu (there was no way for the user to change this
anyway).

Also, removed use of explicit size of hash storing Unicode block names/ranges.