Stop autosave clashing with save
Fixes #196

It was previously possible for an autosave to happen during a Save, Save As, or
Save Copy As. Since some of these do some manipulating of the filenames, this was
potentially a dangerous operation, and because some save operations pop a dialog
for the user to respond to, they could take a long time, increasing the likelihood of
a clash.

This edit turns off autosave during the manual save processes. It also tidies the code,
putting all the main autosave code into the same module, and simplifies the calls to
the main subroutine.

Finally, part of the Save As and Save Copy As code that needed editing (to avoid a
premature return) turned out to be unnecessary, so was removed. For the record,
there was code to check for a Bin File Collision. Historically, filename.txt had a bin
file called filename.bin. Hence, if the HTML file was saved as filename.html, it would
have had the same bin file name. More recently (1.0.25 or before) the naming of the
bin file was changed to be filename.txt.bin and filename.html.bin, so there is no
possibility of a collision.