Stop setting.rc from affecting results of test suite
The setting.rc file was renamed at the start of the tests to avoid it affecting the
test results. However, it was already being loaded before that point.

Now, the global variable to say we're in testing mode is set earlier in guiguts.pl,
and the loading and saving of setting.rc is skipped when in testing mode.

Also, removed the old attempt to save & restore setting.rc.

This also affected the results of two tests:
jeebies runs in "paranoid" mode
CSS validation runs at CSS level 3
so baseline files for these needed updating.

Fixes #233