Retain GC/BL view options when dialog is closed
If the View Options dialog was closed, running gutcheck/bookloupe showed all
errors, rather than hiding the ones the user had selected.
The settings are now used even when the View Options dialog is closed. To ensure
the user knows they have hidden error types, the number of error types hidden is
displayed with the number of queries found.

Fixes #231