Fix error messages when Import is cancelled or bad file loaded
File-->Content Providing-->Import One File with Page Sep Markup had a check to
see if a filename was given before it tried to open it. However, even if none was given
it still went on to execute the rest of the routine. Fixed by returning earlier if
cancelled.

Also, if a junk file is loaded, just abandon the routine, rather than try to work through
with uninitialised variables. Situation too rare to warrant detailed error handling.

Fixes #290