Avoid deleting text when converting inline sidenotes to HTML
Although files should not come out of the proofing rounds with sidenotes inline,
some PPers may make them inline in certain circumstances, e.g.
`This sentence [Sidenote: Notes] has an inline sidenote`
Since the code was assuming the sidenote would have a trailing `</p>` it deleted
4 more characters after the closing `]`.
Now it only does that if they characters are `]</p>`. If it's just a `]`, then just that
one character is deleted.

Fixes #271 and prepares the way for #272 to work successfully if implemented.