Fix errors during Link Check on edited utf8 file
If a file has been edited, then when Link Check is run, the code wrote a special
temporary file for Link Check to work on, in addition to the one that is saved for the
other tools to use. This Link Check tempfile writing didn't handle utf8 properly,
causing `wide character in print` messages.

Now changed to use the same temp file that other tools use.
Also extracted the saving to a temp file into a separate routine.

Fixes #274