Fix failure to update if search too quickly
If user clicks Search twice within a second, the window was not updated due to #153.

Change logic of replaceall() to call searchtext() in silentmode, rather than relying on it
not to update more than once a second. This means when searchtext() is used in
single search mode, it will not suppress updating, even if less than a second since the
last search.

Fixes #305