Make jeebies use same tempfile mechanism as other tools
Jeebies used to save the current file if edited, and run on the current file, unlike
other tools which saved to a temporary file for checking.

Now all tools use the same mechanism, with temp filenames stored in variables.
The temp files are also in the user's working directory, not the installation dir.

Also pptxt no longer reports the filename it is converting - no other tools do it,
and it's a nuisance in the text environment.

Fixes #299