Stop scanno auto-advance affecting `string not found` alert
If stealth scanno auto-advance is turned on, then normal string searching no longer
warns the user with bell and flashing Search button when string is not found.

Main search routine is used for stealth scanno search, but the check for auto-advance
needed to also check that it was a scanno search.