Remove use of yview for mouse wheel and forcing scrollbars
It appears as though BindMouseWheel is not necessary for scrolled widgets to
scroll using the mouse wheel. With or without it, some do and some don't, probably
depending on focus and type of widget being scrolled.

Also, yview was used to scroll down and up one line on several widgets when they
get created, possibly to force optional scrollbars to become active. Without it, as
long as scrollbars are not marked as optional, their behaviour seems identical.

Please check on non-Windows platforms if possible.
Also, does this fix #77 - failed to autoload yview?