Improve Page Separator Fixup undo/redo
Page Separator Fixup had its own undo and redo lists, storing numbers of operations
to be undone, possibly because of the previously recursive nature of the code. It was
not reliable and did not always return the text to its previous state.

Special undo/redo lists have been removed, and the generic undo/redo method
used, including glob functionality around single user operations. Hence it is possible
to undo a single manual join or many auto-joins with a single click.