Check regex validity more thoroughly to avoid warnings
Although a check on regex validity already existed, it only detected compile errors,
not Perl warnings.

In order to trap these it was necessary to override the warning handler. However,
these are only called once for a specific error on a specific regex. It was therefore
necessary to replace the timer checking the regex every 0.5 seconds with a validate
routine which is better anyway. One example of why this is needed is if the dialog
was popped with the same bad regex again - no warning would be raised and the
regex would not be flagged as bad to the user.

This necessitated changing the Text field to an Entry field (which was noted as a
Todo anyway) & also adjust associated method calls.

Fixes #315