Fix obscure infinite loop in Search & Replace
To avoid a bug in Text::Tk FindAndReplaceAll method, there was logic to only call it
when not using regex and replace string does not contain search string (including
case-insensitive version). However, the search term was checked with isvalid, which
is to check for a valid regex. This was needed because Perl regex =~ operator
was used to check if replace string contained search string.

So, if the non-regex string happened to be an invalid regex, e.g. `[d`, the logic failed
and the fast, buggy version was called.

Fixed by not using isvalid (since we know we're not doing a regex search), and
using the index function instead of =~ to check for replacement containing search
string.

Fixes #327