Do not flash status bar label in addition to S&R dialog search button
If bell is turned off, the first label in the status bar is flashed in place of sounding the
bell. In the S&R dialog, the Search button flashes if the search string is not found.
It is not necessary and takes too long to flash both of these.

Suppress the status bar flash in the specific S&R dialog case with an optional arg to
the soundbell routine.