Simplify Word Frequency word list build and avoid wide char messages
Before building the word frequency list, the code either saved the current file, (or
saved to a tempfile), then re-read the file a line at a time to build the word list.
Saving to a tempfile wasn't done in a utf-8 friendly way, leading to warnings.

Instead, just get the lines of the currently loaded file one at a time and build the list
from them. This gives identical results, is simpler and avoids writing the tempfile at
all.

Fixes #310