Improve alphabetical sorting for Word Frequency lists
Sorting was done using words converted to lower case and de-accented. This meant
that words which were identical apart from case/accents would appear in a random
order in the Word Frequency lists.

Now the sorts do the same sort as before, but if the keys are identical when
compared lowercase/de-accented, then they are compared with true case and
accents. So `The` will always precede `the` and `thé`, etc.

This will also aid regression testing in case of future changes to Word Frequency
code since the order of the output is well-defined.

Fixes #334