Tidy up after MR #328
Most text entry fields in GG are Entry widgets, but the ones in the Search & Replace
dialog were historically Text widgets. #328 changed these to Entry widgets.

However, as noted in #97, paste behaviour is inconsistent between Text and Entry
widgets, specifically if the text in the field is selected, then the paste overwrites
the selection only in Text widgets, not Entry widgets.

This commit implements the same paste behaviour for Entry widgets. This is
necessary for the S&R dialog at least, but is actually useful in the many other
Entry widgets. It is standard Windows behaviour and pretty harmless on other
platforms.

Testing also revealed a call to non-existent method. Entry::tagAdd to set a selection
is replaced by Entry::selectionRange.

Finally, it is necessary to catch errors from clipboardGet when pasting. This is done
using eval {block}. This stops error output if pasting when there is either nothing in
the clipboard, or it is not a string. In addition to using eval in the new paste
routine, it has also been added to the main text window paste code which would
otherwise give an error when in Overstrike mode.

Fixes #97
Fixes #338