Stop bookloupe errors reappearing in the error list window
When errors reported by bookloupe are right-clicked, they are removed from the list.
However, if the view options were then changed, removed error lines reappeared
in the list

Oversight when commonising the code for all the different error checks. Because
bookloupe has the View Options, its error list can be refreshed, so the error mark
needs undefining when the error is removed from the list.

Fixes #349