Improve Greek transliteration dialog
1. Use text labels rather than images for the Greek letter buttons. This means the
inserted character in the Greek dialog will match the glyph on the button. Also, some
of the characters had an acute accent instead of a breathing mark in the old version.
This means Greekgifs and GifMaker files are no longer required.
4. If the ou ligature is clicked, it now inserts the ou ligature character. If the user
wants separate o and u, they can just the o and u buttons. This makes more sense
now that utf8 text & HTML versions are preferred.
5. Some bits of code have been commonised by splitting bi-directional conversion
routines with (virtually) no common code, paving the way to pass the conversion
routine as an argument to the common callback routine.
6. Extract a long validation routine to improve readability of the main dialog popup
routine.
7. Don't reset the type of input (Latin-1, name, HTML, utf8) to Latin-1  if dialog is
closed then re-opened. Default to utf8 when program starts.