Fix Greek dialog being zero-sized on Macs
The geometry reported by `$::lglobal{grpop}->geometry;` appears incorrect.
It shows width & height of 50x8 which are the width and height of the scrolled
text widget contained in a frame contained in grpop, when it is expected to
return the pixel width & height of the grpop dialog.

Hopefully will fix #362
Whether this is a bug or not, and whether it is related to the Mac size issue, changing
from using positionhash to geometryhash fixed a similar issue in #204, so it's worth
a try.