Improve HTML pagenum markup
1. Add checkbox to HTML generate dialog to skip all but the last when several
pagenums are in the same place, which happens with blank pages. In general
these do not need page numbers and the PPer has to go through and delete them
after autogeneration.
2. If that toggle is turned on (the default), there will only be one page number within
the pagenum span, so instead of needing an `<a>` element adding just to provide
an anchor so the page can be linked to, just add an id to the pagenum `<span>`
3. Saved by the tests! Needed a tiny change to the regex used when trying to
place chapter headings intelligently round pagenums.