Replace Latin-1 chart with Commonly-Used Characters Chart
Since books are not generally Latin-1, more helpful to PPers to have common
characters, such as curly quotes, fractions, emdashes, etc., available in a chart.
1. Replace Latin-1 chart with Common Characters Chart
2. Mouse-3 copies to clipboard, like the Unicode Chart does
3. Spare buttons at the end are left blank (actually a space) which the user can
configure with Ctrl-Mouse-1
4. Config dialog accepts a single character, or a decimal ordinal, or a hex ordinal.
Prefix the hex ordinal with `0x`, `\x` or `x`, or if it has letters [a-f] in it, then prefix
may be omitted. For any other character string, just take the first character.
5. User-defined characters are stored between runs of the program, and in dialog
is upgraded in a future release, they should be preserved to fit into any unused
spaced in the new version.
6. Make a couple of helper functions common between new dialog and Unicode
dialog