Fix S&R bug when text substituted for $1 contains literal $2
In order to facilitate GG's regex extensions such as `\C...\E` and `\G...\E`, some aspects
of the substitution are done in parts rather than as a single perl regex substitution.
As part of this, parenthesis grouping/capturing and substituting were done for one
group at a time. The problem occurred when the substitution for `$1` contained the
literal text `$2`, e.g. `$234.56`. Then later, the `$2` was substituted with the text that
the second paren pair captured.

Fixed by escaping dollar characters in the match strings, so the substitution for
`$1` will contain `\$2` and won't be further substituted. Of course, a `$2` in the replace
term is still replaced with the second group match string.

This took quite a long time to determine where the problem was arising and work
where the escaping was required. I believe it fixes the issue, and I haven't been able
to make other more complex regex replacements go wrong, even if they include $2
combined with extensions such as `\U...\E`

Code also made less repetitive by using an array to capture the matches rather than
8 scalar variables

Also, removal of superfluous and potentially confusing check on regex option
inside an `unless` that precludes regex matching.