Save word count threshold for marked up phrases
In the Word Frequency dialog, the Ital/Bold/SC button checks for the same phrase
occurring with and without markup. A configuration dialog is also popped that sets
the maximum number of words in such a phrase. Previously this was set to 4 on each
run of the program. Now save the value in the setting.rc file. Also, consider 0 to mean
no limit, rather than a limit of zero words which is not useful.

Also allow config dialog to be popped from the Prefs menu where other preferences
are set, and remove a couple of "beta" labels on menu options, since they have been
in "beta" for 10 years.