Provide a default compose key for Mac users
Define the default compose key to be Ctrl-m.
This is currently defined to indent by 1 space, but Alt-RightArrow does the same
action, so Mac users can use that.
Move the compose key binding to be the last binding defined in case the user wants
to reuse a key for compose that is previously defined as something else.
Now that the argument to keybind() could be a user-defined key, ensure that the
key letter is lowercase - necessary because of the odd behaviour of key bindings
when caps lock is on, etc., which keybind() has to deal with.
Add Alt left & right arrow to the accelerators displayed on the indent buttons to
support Mac users who can potentially no longer use Ctrl+m.