Fix bad text index error during sidenote fixup
Looks like the end point of a delete was omitted, possibly always like this, but it
only executes if the sidenote doesn't have a blank line before it, which they generally
should have.
Additional safety check also added - only necessary if sidenote is on the first line of
the file.

Fixes #236