Fix error pasting certain unicode strings
Fixes #179
Certain unicode Greek strings were pasted either correctly or as Latin-1 garbage
depending on apparently random minor changes such as whether a terminating
end of line character was included. Two newlines followed by a lowercase iota
failed, but with just one newline preceding the iota it succeeded.
The bug appeared to occur within Tk::Text::clipboardPaste, possibly to do with the
use of Tk::catch or local $@. If the equivalent code is used instead of a call to
clipboardPaste, all the unicode strings appear to paste correctly.

Also, fixed a minor bug when selecting all text in a file (via Ctrl+a) and then pasting
some new text. The old text was deleted, but the first line of the new text was lost.
This bug is in the original TextUndo::ReplaceSelectionsWith, and also in the
overridden version used in GG.