Ensure only one plus sign is added to HTML Markup dialog buttons
When adding class/attributes to an HTML Markup dialog button a plus sign is
appended. It was possible to add 2 or more plus signs, by editing the configuration
further.

Now only one plus sign is ever added, and all plus signs are removed if the
class/attributes field is cleared