Improve handling of busy in spellcheck dialog
Use of the Busy method was a bit haphazard. As well as changing the cursor so the
user knows something is happening, Busy causes button and key input to be ignored.

Extract into Utilities routines the calls to Busy and Unbusy from the top widget with
recurse option to include all lower widgets. Use these routines at the start and end
of each user action where these have keyboard shortcuts. This greatly reduces the
possibility of the user holding down a key and the repeated keystrokes causing a
"deep recursion" error or crash.

Fixes #181 (or at least makes it hard to get the error and I haven't been able to crash
the program).