Improve display of warnings and errors
Warnings and errors whether issued by Tk or via warn/die in the application code
were only output to stderr, which might not be noticed by the user until it was too
later, or which on some platforms would vanish when they exited the program.

1. Override the handling of warn and die to output to a message log dialog as
well as to stderr.
2. Sound the bell (or flash the warning icon) if a warning or error is output and pop
the message log to the front.
3. Allow the user to pop the message log via the Help menu.
4. In the case of die actually going to terminate the program, give the user a chance
to copy/paste messages first.
5.. Tidy up a few places that used warn/die in the code.