Add best practices code for floated images to HTML header
The Best Practices document suggests making the figleft and figright classes behave
like figcenter on handheld devices.
A common alternative is to force the images to float, overriding ebookmaker's
default behaviour.

Provide both options, with the BP version uncommented, so the PPer can choose
which is most appropriate for the book and image sizes used.