Stop user removing errorcheck entry while already removing one
If the user right-clicked twice in the errorcheck box in rapid succession, then the
system might be refreshing the window when the second click was processed,
leading to an error message.

To avoid this, use busy() to ignore further clicks while first click is being processed.

There was a call to the `after` method in the code, which sleeps for a short period
ignoring clicks - this was perhaps a previous attempt to avoid the same issue, but
has proved insufficient since recent alterations to dialog updating.