Fix crash during pasting
MR #388 to fix #179 changed the routines used for pasting into the main window.
The new method caused crashes with some moderate/large amounts of text.

Revert to the old method for default pasting, but allow user to choose the
alternative method, either from the Edit menu or with Ctrl+Alt+v