Preserve underscores when creating ids and anchors
This is because they are particularly useful to preserve when a filename is being
converted to an id for an HTML image. It's harmless and probably won't happen
apart from that, e.g. for the anchor used in the automatically generated ToC,
because books don't generally have underscores (they are used in the text version
for italics, but not in the HTML version).