Reduce default width of SR dialog fields
Since recent font changes, SR dialog can be wide. It has never had the capability to
store its width, so when user resizes, it returns to its default size.

A quick fix to store the width using the geometryhash will not suffice, since resizing
is complex for this dialog as the number of replace rows can grow and shrink via
the code, and the width needs to be adjustable by the user.

#460 remains as a long-standing issue to be addressed, but not in a bug fix release,
so reducing the default width from 60 characters to 40 is the best option for now.