Fix placement of two closing footnote divs
Two bugs:
1. If a pagemarker directly followed a footnote, the closing div for the footnote would
be inserted after the marker, meaning the page label would come in the footnote
instead of at the top of the next page - long-standing bug.
2. The code to search for the end of the block of footnotes was broken by a 1.1 edit
to improve the look of the HTML code by putting a blank line between the closing
`</p>` and the closing `</div>` for each footnote. Bug in 1.1.

Bug 1 is fixed by inserting the closing div before deleting the closing ]
Bug 2 is fixed by making the search more flexible to allow for blank line at end of
footnote.