Fix uninitialized value error when saving file with path containing quote
If the path to a file contained a quote, the quote is escaped when the "file opened"
operation is being written to the list of operations stored in the bin file. However,
the escaped version of the hash key was also used for looking up the timestamp,
which was then undefined because the non-escaped hash key was used when storing
the timestamp originally. Error only occurred when operations history was enabled.

Also, removed a redundant escaping of the operation name in operationadd()
which was done after the operation variable had been used.

Fixes #476