Handle multi-line search strings better
Change from Text to Entry field in the S&R dialog meant that instead of being
restricted to just the first line of a multi-line search string, a long string of
concatenated lines was used. Since this could be very long, it could break the
search history pulldown menu.

Fixes #495