Grab focus back after Image Viewer is spawned
It isn't possible to avoid the Image Viewer getting focus, so the only solution is to
grab it back. Rather than wait 1/5 second which means the Image Viewer has focus
for a noticeable length of time, try every 1/100 second for 1/5 second. This makes
the Viewer only flash briefly with focus. If the user's computer is too slow and the
Image Viewer has not been given focus within the grab period, then GG will lose
focus (once the viewer is running) as happened prior to this fix.

I considered a configurable grab period, but it's not worth the code and user
interface complexity.