Make Word Frequency Ligatures respect Suspects Only flag
The accented character check (and others) in Word Frequency take note of the
Suspects Only flag, in which case they only report if the word appears in both the
accented and non-accented form.

The Ligature check was just displaying all words containing ligatures or potential
ligatures, i.e. ae and oe, regardless of the Suspects Only flag.

Setting of Suspects Only is now respected, and if enabled, words are only reported
if they occur in both ligatured and non-ligatured form.