Catch all footnotes without closing brackets
Under some circumstances, footnotes without closing brackets were not flagged as
errors during First Pass, leading to errors later when the user tried to move them.

If the code didn't find a suitable end, it already set the endpoint to be 10 characters
after the start (which is physically impossible), so detect this case and flag it as a
potentially long, potentially missing bracket case.