Make `$t` substitute correctly for utf8 characters in Custom menu (#523)
The system command under Windows wasn't capable of handling utf8 characters
which meant that the Custom menu couldn't look up a word in a dictionary if the
word had accented characters.

Fixed by using the Win32::Unicode module which implements a systemW equivalent
to system. This is only required for Windows, and the cpan installer script now has a
section for Windows-specific modules.