Make undo work correctly after S&R inserts string terminating in newline
A bug in the Tk::TextUndo method insert_UNDO causes an undo to delete too much
when a replaced multi-line string ends in a newline character.

Override the method by including it in TextUnicode and adjusting the method of
calculating the length of the residue of the string after the final newline.