Only convert correctly formatted index page numbers to links
Page numbers are required to have a comma and space preceding them when an
index comes out of the rounds. This was not enforced by the code, so some numbers
got converted that shouldn't be, e.g. in
`The Taking of Pelham 123, 7, 19, 26.`
the 123 would have been converted.
Also, if a 4 digit number appeared (like a date or a genuine page number at the
end of a list), the first or last 3 digits could sometimes be converted.

Rules now (fairly) strictly enforced. This is preferable, because if the index is badly
formatted, it is better for the PPer to be alerted by the link not being created than
to silently and possibly wrongly attempt a conversion anyway. A formatting error
like that would quite likely need some correction in the text file anyway, so it's
good to be aware of it.