Make Search and Replace dialog resizable in width
The height of the S&R dialog needs to change depending on if it is in single, multi
or scanno mode - this is under program control. Also, the user should be able to
set the width. Particularly since the user can set their font easily, some have found
that the default field width in characters causes the dialog to be too wide, while
others have found that the width is too small.

The mechanism used by other dialogs is to save the width and height in pixels
whenever the dialog is resized and to pop it at the same size next time. However,
it is not possible to do this for width only, and once the height has been set once
the dialog will not grow or shrink programatically when switched between modes.

Solution is to measure the height of the widgets being placed in the dialog and
set the dialog height to be the total of these. This needs to be updated whenever
the mode changes and when the font is changed.

Also the order of widget packing has been adjusted to ensure the search history
pulldown buttons are not clipped off the side of the dialog when it is shrunk.