Stop bookloupe saying it won't report things that it will
Bookloupe does an early check and, for example, if it finds a lot of long lines, it says
it won't bother reporting them. However, with the verbose flag on, which it always is
when run through Guiguts, it then reports everything, including the things it said it
wouldn't.

This commit skips over those "Not reporting..." messages, but doesn't change what
is actually reported.