Preserve blank lines after block markup during Page Sep Fixup
If block markup continued through a page boundary, and was followed by 2 or more
blank lines, these were reduced to 1 blank line, e.g.
```
...
*/
---File: --------page sep
/*

Section head in ToC, for example
...
```

Code was only checking for 1 blank line and assuming it was the start of a
paragraph. Changed to check for 4 or 2 first and treat accordingly.