Join footnotes correctly after user has made manual edits to file
If the user runs Footnote First Pass, then manually edits the file before attempting to
Join Footnote to Previous, the joining fails. This is because the code stores indexes to
the line and column where the footnotes start. These indexes are shown to user in the Footnote List dialog, but aren't safe to be used for manipulation of footnotes
when the user has edited the text.

Switch to using the permanent footnote marks in the joining routine.

Also, exit the routine immediately if user hasn't chosen at least the second footnote
to join to the previous.