Use recommended method to get Unicode character names
Guiguts has always read Unicode character names from `lib/unicore/Name.pl`.

In Perl 5.32, the format of this file changed, and its header now has the warning
"INTERNAL PERL USE ONLY" and that the format and existence of the file could
change in future Perl versions. The recommended method is to use the core
Unicode::UCD module, and sample code is supplied in the module documentation.

Use the relevant parts of the sample code - the only other change being that the
ordinals are now decimal integers rather than hex strings.

Also, when popping dialog, put the focus in the text field the user has to type into.