Fix HTML page numbering errors
1. If two page markers were adjacent, in one place the first one encountered might
be used and in the other the highest one numerically might be used, leading to an
id of 'Page_292' on page '293', for example. Now always choose the highest
numbered.

2. If Roman page xiv immediately preceded Arabic page 3, it treated the Roman
number as the "last" since it was a bigger number. Since Roman page numbers
nearly always precede Arabic, it's safer to force that than the reverse.

3. Minor fix to baseline bookloupe test file while tests being edited.