Pass title and author to locally-run ebookmaker
Extract title and author from `<title>` element in HTML file and pass as arguments
to ebookmaker. This causes the epub/Kindle files to be named with the book title
and the title and author are also stored in the files, so they show correctly on the
reader's e-reading device. This is helpful for creating smooth-reading versions.