Convert all accented characters during "display" conversion
If an accented character was not handled by the language-specific conversions for
display purposes, then only the Latin-1 supplement characters were also converted.

Changed to include all Latin characters from the Extended and Additional blocks.
Also ordered and split list of converted characters in the Latin-1 Supplement block
to match the handling of the other blocks.

Fixes #569