Avoid interpreting `I/`, etc. in text as markup
There were several places where 'I/' and potentially 'L/' and similar strings in the
text of the book could be wrongly interpreted as markup. These affected both text
rewrapping and HTML conversion. Various tests have been tightened up, such as
insisting `I/` is on a line on its own, or possibly with [n.n,n] rewrap directive.

Note that it is still possible that such a string in the text could end up on a line on
its own and be misinterpreted, but very unlikely. `I/J` or 'I/27' might occur, but rarely
`I/` with no punctuation or letters after it.

Fixes #564