Preserve page label positions during ASCII table operations
The operations Move Left/Right (with rewrap on ), Auto Columns, Step to Grid and
Grid to Step all caused any page markers within or immediately following the table
to be moved to the beginning of the table.

Now preserved by storing the current positions and adjusting them. Due to the
reorganisation involved when converting grid to step format, it is necessary to flag
the positions using a temporary character, which is removed later.

Fixes #561